app/api/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
python-socketio